requirements.txt
torch==1.10.1
dm-control==0.0.403778684
dm2gym==0.2.0
opencv-python==4.5.5.62
realworldrl-suite==1.0
hydra-core==0.11.3
termcolor>=1.1.0
tensorboard>=2.4.0
imageio>=2.9.0
numpy>=1.19.1
matplotlib>=3.3.1
gym==0.17.2
jupyter>=1.0.0
pytest>=6.0.1
sk-video>=1.1.10
tqdm>=4.48.2
numba==0.56.3
filterpy==1.4.5
